style: bring the motion sources in line with the project's lint rules
The native preset lints headers as well as sources, so every header the
imported motion tree brings in was being checked and none of it passed.
Around 520 findings, the bulk of them one rule: .clang-tidy asks for
camelBack constants, and RepRapFirmware writes them CamelCase.

Renamed to the project's convention rather than the other way round, so
MaxAxes becomes maxAxes, StepClockRate becomes stepClockRate, and
debugPrintf becomes DebugPrintf. clang-tidy applied its own fixits, which
renames each declaration and every use together; the collisions it could
not see were XYZ_AXES and XyzAxes both mapping to xyzAxes, a test-local
xAxis that started shadowing the global one, and the identifiers inside
preprocessor-disabled blocks that it never parsed.

The cost is that Movement/ is no longer a clean diff against upstream
RRF - re-syncing it is now a merge. The headers that claimed otherwise
have been corrected rather than left to mislead.

The rest were fixed in code: [[nodiscard]] on the accessors, member
initialisers, named parameters, a byte-wide MessageType, and a pointer
rather than a reference member in the GCodes shim. RepRapShim's
singletons became ordinary members, because static accessors reached
through `reprap.` are themselves a lint error at every call site - and
`reprap` is the singleton, so what hangs off it need not be.

+++ src/DuetSbcInterface/src/Compat/Platform/Platform.h
@@ -17,7 +17,7 @@
 // The subset of the firmware's MessageType that reaches this code. The values do not have to match
 // the firmware's: nothing on this side puts them on the wire, and DCS classifies messages itself
 // from the LinkEvents log level.
-enum MessageType : uint32_t
+enum MessageType : uint8_t
 {
 	DebugMessage = 0,
 	GenericMessage,
